Delete BackgroundTile.cs
Instead of creating one class per tile type, I'm going to try to use the
same class for all of them and only allow the user to set specific
properties of the tile based on the selected tile type, and then
save/load that specific set of properties. This should be simpler than
dealing with multiple classes or possibly polymorphism.