Remove unused distinction between tile types.
Once I started adding this functionality on the game rather than the
level editor it quickly occurred to me that I didn't need this. The
foreground and background tiles can and should occupy the same space and
won't conflict with one another. This was less obvious before because
the map makes it look like some background tiles are "behind" foreground
tiles. This isn't necessarily the case; my sprite sheet is just lacking
some tiles.